+++ instrumentation/java-concurrent/java-concurrent.gradle
@@ -5,39 +5,9 @@ ext {
 }
 
 apply from: "$rootDir/gradle/instrumentation.gradle"
-apply from: "$rootDir/gradle/test-with-scala.gradle"
-apply plugin: 'org.unbroken-dome.test-sets'
 
-// This won't work until the akka and scala integrations are split into separate projects.
-//muzzle {
-//  pass {
-//    coreJdk()
-//  }
-//}
-
-testSets {
-  slickTest {
-    filter {
-      // this is needed because "test.dependsOn slickTest", and so without this,
-      // running a single test in the default test set will fail
-      setFailOnNoMatchingTests(false)
-    }
+muzzle {
+  pass {
+    coreJdk()
   }
 }
-
-compileSlickTestGroovy {
-  classpath += files(sourceSets.slickTest.scala.classesDirectory)
-}
-
-dependencies {
-  // This is needed for Scala ForJoinTask/Pool instrumentation
-  compileOnly deps.scala
-
-  slickTestImplementation project(':instrumentation:jdbc')
-  slickTestImplementation deps.scala
-  slickTestImplementation group: 'com.typesafe.slick', name: 'slick_2.11', version: '3.2.0'
-  slickTestImplementation group: 'com.h2database', name: 'h2', version: '1.4.197'
-}
-
-// Run Slick library tests along with the rest of unit tests
-test.dependsOn slickTest
