--- instrumentation/java-concurrent/scala-testing/scala-testing.gradle
@@ -1,9 +1,0 @@
-ext.skipPublish = true
-apply from: "$rootDir/gradle/instrumentation.gradle"
-apply from: "$rootDir/gradle/test-with-scala.gradle"
-
-dependencies {
-  compileOnly deps.scala
-
-  testImplementation deps.scala
-}
