+++ settings.gradle
@@ -98,7 +98,6 @@ include ':instrumentation:java-classloader:osgi-testing'
 include ':instrumentation:java-classloader:tomcat-testing'
 include ':instrumentation:java-concurrent'
 include ':instrumentation:java-concurrent:kotlin-testing'
-include ':instrumentation:java-concurrent:scala-testing'
 include ':instrumentation:jaxrs:jaxrs-1.0'
 include ':instrumentation:jaxrs:jaxrs-2.0'
 include ':instrumentation:jaxrs:jaxrs-2.0:jaxrs-2.0-jersey-2.0'
@@ -148,6 +147,8 @@ include ':instrumentation:rediscala-1.8'
 include ':instrumentation:redisson-3.0'
 include ':instrumentation:rmi'
 include ':instrumentation:rxjava-1.0'
+include ':instrumentation:rxjava-1.0'
+include ':instrumentation:scala-concurrent'
 include ':instrumentation:servlet:glassfish-testing'
 include ':instrumentation:servlet:servlet-common'
 include ':instrumentation:servlet:servlet-2.2'
